# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "description": "APIs for Post Away application",
    "title": "PostAway API"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "in": "cookie",
        "name": "jwtToken",
        "type": "apiKey"
      }
    }
  },
  "paths": {
...
    "/api/posts": {
      "post": {
        "tags": ["Posts"],
        "summary": "Add a post",
        "description": "Add a post to the system. This endpoint accepts a multipart/form-data request with an image file and a caption. The `imageUrl` and `caption` are two form fields expected in the request body. The user ID is obtained from the authenticated user's session.",
        "security": [{ "JWT": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "imageUrl": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file for the post"
                  },
                  "caption": {
                    "type": "string",
                    "description": "Caption for the post"
                  }
                },
                "required": ["imageUrl", "caption"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post added successfully"
          },
          "400": {
            "description": "can't create the post"
          },
          "500": {
            "description": "creating new post failed"
          }
        }
      },
      "get": {
        "tags": ["Posts"],
        "summary": "Get all posts by user",
        "description": "Retrieves all posts created by the authenticated user. The user ID is extracted from the authenticated user's session.",
        "security": [{ "JWT": [] }],
        "responses": {
          "200": {
            "description": "Posts fetched successfully"
          },
          "404": {
            "description": "Posts not found"
          },
          "500": {
            "description": "getting all users post failed"
          }
        }
      }
    },
...
  }
}